anthonytoyco - fixed day 0 solution output

+++ advanced/days/0/solutions/anthonytoyco/anthonytoyco.py
@@ -3,5 +3,5 @@ import math
 a = 2546375247
 b = 45987482082
 
-print((a * b) / 2)
-print(a + b + math.sqrt((a**2) + (b**2)))
+print(int((a * b) / 2))
+print(int(a + b + math.sqrt((a**2) + (b**2))))
